Created stories component
